requirements.txt
# Core dependencies
flask==2.3.3
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0

# LLM and API
openai==1.3.5
requests==2.31.0

# Utilities
structlog==23.2.0
colorama==0.4.6
pydantic==2.5.2
pydantic-settings==2.1.0